+++ web-storage-setter/src/contentScript.js
@@ -1,2 +1,0 @@
-
-//# sourceMappingURL=contentScript.js.map